Feature gate verify
