+++ Cargo.toml
@@ -14,7 +14,9 @@ ed25519-dalek = { version = "1.0.0-pre.1", optional = true }
 [features]
 default = ["std"]
 std = [
-  "ed25519-dalek",
   "parity-codec/std",
+]
+verify = [
+  "ed25519-dalek",
   "schnorrkel",
 ]
